+++ src/main.py
@@ -12,10 +12,21 @@ from contextlib import asynccontextmanager
 from src.db.database import SessionLocal
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
+from alembic import command
+from alembic.config import Config
+from src.db.database import engine
+from sqlalchemy import inspect
 
+def check_if_table_exist():
+    inspector = inspect(engine)
+    tables = inspector.get_table_names()
+    return bool(tables)
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
+    if not check_if_table_exist():
+        alembic_cfg = Config("alembic.ini")
+        command.upgrade(alembic_cfg, "head")
     db = SessionLocal()
     try:
         admin = User(
